backend

+++ clean.py
@@ -4,7 +4,7 @@ here = os.path.dirname(os.path.abspath(__file__))
 files = os.listdir(here)
 for i in files:
     if 'output_' in i:
-        if '53d' not in i and '41c' not in i:
+        if '53d' not in i and '41c' not in i and 'a7e' not in i:
             shutil.rmtree(i)
     if '__py' in i:
         shutil.rmtree(i)
